# renderer: add non-bitwise alternative for u_Color and u_ColorModulateColorGen