--- a/src/engine/renderer/glsl_source/forwardLighting_vp.glsl
+++ b/src/engine/renderer/glsl_source/forwardLighting_vp.glsl
@@ -32,8 +32,8 @@ uniform mat4		u_LightAttenuationMatrix;
 uniform mat4		u_ModelMatrix;
 uniform mat4		u_ModelViewProjectionMatrix;
 
-uniform uint u_ColorModulateColorGen;
-uniform uint u_Color;
+uniform colorModulatePack u_ColorModulateColorGen;
+uniform colorPack u_Color;
 
 uniform float		u_Time;
 
@@ -64,7 +64,7 @@ void	main()
 	VertexFetch( position, LB, color, texCoord, lmCoord);
 
 	// assign color
-    color = color * ColorModulateToColor( u_ColorModulateColorGen ) + unpackUnorm4x8( u_Color );
+	ColorModulateColor( u_ColorModulateColorGen, u_Color, color );
 
 	DeformVertex( position,
 		      LB.normal,