--- a/src/engine/renderer/glsl_source/generic_vp.glsl
+++ b/src/engine/renderer/glsl_source/generic_vp.glsl
@@ -36,8 +36,9 @@ uniform vec3 u_ViewOrigin;
 
 uniform float u_Time;
 
-uniform uint u_ColorModulateColorGen;
-uniform uint u_Color;
+uniform colorModulatePack u_ColorModulateColorGen;
+uniform colorPack u_Color;
+
 
 #if defined(USE_TCGEN_ENVIRONMENT)
 	uniform mat4 u_ModelMatrix;
@@ -63,10 +64,7 @@ void main() {
 	vec2 texCoord, lmCoord;
 
 	VertexFetch( position, LB, color, texCoord, lmCoord );
-	float lightFactor = ColorModulateToLightFactor( u_ColorModulateColorGen );
-	color.a = ColorModulateToVertexColor( u_ColorModulateColorGen ) ? 1.0 : color.a;
-	color = color * ColorModulateToColor( u_ColorModulateColorGen, lightFactor )
-		+ unpackUnorm4x8( u_Color ) * vec4( lightFactor, lightFactor, lightFactor, 1.0 );
+	ColorModulateColor_lightFactor( u_ColorModulateColorGen, u_Color, color );
 
 	DeformVertex( position, LB.normal, texCoord, color, u_Time );