--- a/src/engine/renderer/glsl_source/shadowFill_vp.glsl
+++ b/src/engine/renderer/glsl_source/shadowFill_vp.glsl
@@ -22,11 +22,12 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 /* shadowFill_vp.glsl */
 
+#insert common
 #insert vertexSimple_vp
 #insert vertexSkinning_vp
 #insert vertexAnimation_vp
 
-uniform uint		u_Color;
+uniform colorPack u_Color;
 
 uniform mat3x2		u_TextureMatrix;
 uniform mat4		u_ModelMatrix;
@@ -73,5 +74,5 @@ void	main()
 	var_TexCoords = (u_TextureMatrix * vec3(texCoord, 1.0)).st;
 
 	// assign color
-	var_Color = unpackUnorm4x8( u_Color );
+	var_Color = UnpackColor( u_Color );
 }