… src/engine/renderer/gl_shader.cpp
… src/engine/renderer/tr_backend.cpp
… src/engine/renderer/tr_shade.cpp
… src/engine/sys/sdl_glimp.cpp