Add non-bitwise alternative for `u_Color` and `u_ColorModulateColorGen`, it is when GLSL 1.20 isn't supported.

The material system assumes GLSL 1.20 is supported.

Written on top of:

- https://github.com/DaemonEngine/Daemon/pull/1576
- https://github.com/DaemonEngine/Daemon/pull/1577
- https://github.com/DaemonEngine/Daemon/pull/1578