@illwieckz 2025-03-01
Context: I noticed that since 0.55.2 we broke GL 2.1 (GLSL 1.20) compatibility.

What broke compatibility was the usage of `uint` and bitwise operations that starts to be supported only with GL 3. (GLSL 1.30).

I don't mind if we use bitwise operations and uint in any feature that is not part of `low` and `lowest` presets.

So I ported the `u_Color` and `u_ColorModulate` code to emulate both `uint` and bitwise operations on GLSL 1.20.

I have noticed that writing a `GLUniform1ui` when `uint` isn't supported leads to a segfault, so I use `GLUniform1i` to store the bitfield to keep the code unique on C++ size. Both `int` and int32` having the same size, it works as the data is properly carried anyway.

I don't plan to do this for any other code that is not required by `low` and `lowest` presets. If it happens a not-`low*` preset uses uint or bitwise operations we can disable the feature on GLSL 1.20.

That also means that any feature from some advanced rendering code path (like the bindless one) can rely on `uint` and bitwise operations as much as we want.