@illwieckz 2025-03-01
It's possible to test the code on non-GL 2.1 hardware when running Mesa and setting those environment variables:

```sh
export MESA_GL_VERSION_OVERRIDE='2.1'
export MESA_GLSL_VERSION_OVERRIDE='120'
```