@slipher 2025-03-01
Is https://github.com/DaemonEngine/Daemon/pull/1543#discussion_r1946532883 saying 1.20 can use bitwise operators not true?