@VReaperV 2025-03-01
> Is [#1543 (comment)](https://github.com/DaemonEngine/Daemon/pull/1543#discussion_r1946532883) saying 1.20 can use bitwise operators not true?

The spec for version 120 has bitwise operators, though it does not support `uint` as a type.