@VReaperV 2025-03-01
> Speaking of those bitwise operations, on master I've been getting compile errors again with the Apple Silicon GPU driver because it doesn't like if you `&` signed with unsigned operands or something. I wonder if this can work around it :)

Probably so. Some drivers fail to do the implicit conversions.