@illwieckz 2025-03-01
The GLSL 1.30 version doesn't say bitwise operators are reserved:

- https://registry.khronos.org/OpenGL/specs/gl/GLSLangSpec.1.30.pdf