@VReaperV 2025-03-01
> > > Is [#1543 (comment)](https://github.com/DaemonEngine/Daemon/pull/1543#discussion_r1946532883) saying 1.20 can use bitwise operators not true?
> > 
> > 
> > The spec for version 120 has bitwise operators, though it does not support `uint` as a type.
> 
> What? I just tested… and it works, with the original code, just by using `int` instead of `uint`…
> 
> But that document says:
> 
> * https://registry.khronos.org/OpenGL/specs/gl/GLSLangSpec.1.20.pdf
> 
> > The OpenGL Shading Language has the following operators.
> > Those marked reserved are illegal.
> > bit-wise shift (reserved) << >>
> 
> Actually it's funny because they say `%` is reserved and then illegal too. Though for this one we can just use `mod()` to be safe.

Oh, I was looking at the pre-processor one. But if it works, then whatever I guess.