@illwieckz 2025-03-02
I remember we had to do a non-bitwise fix in the past:

- https://github.com/DaemonEngine/Daemon/commit/df6b609d4c38b55f178cf24d250f99ec5fe602f6
- https://github.com/DaemonEngine/Daemon/issues/490

It looks like the extension we look for is `EXT_gpu_shader4`:

- https://registry.khronos.org/OpenGL/extensions/EXT/EXT_gpu_shader4.txt

> - Integer bitwise operators are now enabled.

We currently load `EXT_gpu_shader4` when available, so me currently testing GL 2.1 on a 4.6 hardware explains why bitwise operations are working.

> Though it should also allow `uint` but I don't get it:

> * Full signed integer and unsigned integer support in the OpenGL Shading Language: