@illwieckz 2025-03-02
So yes with `MESA_EXTENSION_OVERRIDE='-GL_EXT_gpu_shader4'` I get:

```
error: operator '%' is reserved in GLSL 1.20 (GLSL 1.30 or GLSL ES 3.00 required)
```

So replaced `x % y` with `mod(x, y)`, but it looks like that in GLSL those two operations are not that equivalent, because my UI is currently pitch black, but the whole map look properly rendered. The translucency of 2D works, but colors are all pitch black.

At this point I can't exclude it may uncovers a dormant bug in the 2D code.

![unvanquished_2025-03-02_053018_000](https://github.com/user-attachments/assets/e8bae0f3-2dc6-4a5a-bdc7-4716df38165e)