@slipher 2025-03-02
Then the question would be whether Mesa implements the extension on that hardware.