@illwieckz 2025-03-02
> Then the question would be whether Mesa implements the extension on that hardware.

Mesa doesn't, I've checked my previous logs, this extension isn't provided.