@illwieckz 2025-03-02
So, the function that turns the UI black when I enable the alternative code is `vec4 ColorModulateToColor( const in int colorMod, const in float lightFactor )`, which has no reason to be used for UI.