@illwieckz 2025-03-02
I foud the cause of the black UI. The GLSL 1.20 specification says:

> Because of their intended (limited) purpose, integers are limited to 16 bits of precision, plus a sign representation in both the vertex and fragment languages

I had moved the bits for `lightFactor` at the end at the 32-bit integer to guarantee there is no higher bits used…

If I move the bits for `lightFactor` at the end of the 16-bit integer, it works again.

It is said the sign representation is not part of those 16-bits, so it means the code can be purged from my sign hack, and it actually bumps the amount of bits to kinda 17-bit with a bit recoverable with an explicit sign test that would not pollute any division like a stray upper bit can do.

It also means we should encode colorgen on 16-bits instead of 32-bits on GLSL 1.20, which is fine for a `low` preset. The good news is that I first implemented 16-bit packing after doing 32-bit-with-sign-hack. So I just have to recover that code from an old branch.

Being limited to 16-bit plus sign means we still have 6 bits available, plus the sign, (so 7 kinda-bits) to encode future information.