@illwieckz 2025-03-03
So, the float alternative is fully implemented. The bitwise emulation is deleted. Performance is as expected.

||plat23 spectator|plat23 `1920 1920 0 0 0`|0.55.2 benchmark demo
|-|-|-|-
|0.55.2|116 fps|46 fps|6204 frames, 126.3s: 49.1 fps
|this branch|116 fps|44 fps|6204 frames, 112.1s: 55.4 fps

The float packing is different than the one that existed before the uint packing:

Before:

```
colorMod[ 0 ]: red * f
colorMod[ 1 ]: green * f
colorMod[ 2 ]: blue * f
colorMod[ 4 ]: alpha * f
```

```
colorMod[ 0 ]: color * f
colorMod[ 1 ] absolute value: lightFactor
colorMod[ 1 ] minus sign: color += lightFactor
colorMod[ 2 ]: alpha = 1
colorMod[ 3 ]: alpha * f
```

The fact we store the rgb color in a single float means we can store much more data in a single vec4, like the lightFactor, and other things. We can actually store more information with sign and magic values in `colorMod[ 2 ]`. Or storing an index in `colorMod[ 0 ]` and `colorMod[ 3 ]` (like the `uint` packing does) will also make the signs available for storage of more data.